Split stores into subdirectories to satisfy max-lines-per-function linting rule
Each store (items, workspaces, interface) is now a directory with an index.ts
that composes logic from focused sub-files (initialization, validation,
accessors, mutations, etc.), keeping every function under the 50-line limit.

Closes #70
