+++ src/stores/items/initialization.ts
@@ -1,0 +1,12 @@
+import { getAllItems } from "@/db"
+import type { Item } from "@/types"
+import type { Ref } from "vue"
+
+export function useItemsInitialization(items: Ref<Record<string, Item>>) {
+  async function initializeItems() {
+    const storedItems = await getAllItems()
+    Object.entries(storedItems).forEach(([id, item]) => items.value[id] = item)
+  }
+
+  return { initializeItems }
+}
